# `for_each_contiguous_slice` tweaks and optimization