--- a/runtime/buffer.h
+++ b/runtime/buffer.h
@@ -459,30 +459,17 @@ void make_for_each_contiguous_slice_dims(
   next->extent_here = slice_extent;
 }
 
-template <typename F, std::size_t NumBufs, std::size_t... Indices>
-inline void call_fn(const F& f, index_t slice_extent, const std::array<void*, NumBufs>& bases, std::index_sequence<Indices...>) {
-  static_assert(sizeof...(Indices) == NumBufs);
-  f(slice_extent, bases[Indices]...);
-}
-
-template <typename F, std::size_t NumBufs>
-inline void call_fn(const F& f, index_t slice_extent, const std::array<void*, NumBufs>& bases) {
-  call_fn<F, NumBufs>(f, slice_extent, bases, std::make_index_sequence<NumBufs>());
-}
-
 template <typename F, std::size_t NumBufs>
 void for_each_contiguous_slice_impl(std::array<void*, NumBufs> bases,
     const for_each_contiguous_slice_dim* slice_dim, dim_or_stride* dims, const F& f) {
   if (slice_dim->impl == for_each_contiguous_slice_dim::call_f) {
-
-    call_fn<F, NumBufs>(f, slice_dim->extent_here, bases);
-
+    std::apply(f, std::tuple_cat(std::make_tuple(slice_dim->extent_here), bases));
   } else if (slice_dim->impl == for_each_contiguous_slice_dim::loop_linear) {
     const auto* next = slice_dim + 1;
     if (next->impl == for_each_contiguous_slice_dim::call_f) {
       // If the next step is to call f, do that eagerly here to avoid an extra call.
       for (index_t i = 0; i < slice_dim->extent_here; ++i) {
-        call_fn<F, NumBufs>(f, next->extent_here, bases);
+        std::apply(f, std::tuple_cat(std::make_tuple(next->extent_here), bases));
         for (std::size_t n = 0; n < NumBufs; n++) {
           bases[n] = offset_bytes(bases[n], dims[n].stride);
         }