- Split `for_each_contiguous_slice_dim` structure into two structures to save a little stack space and simplify the code
- Let the compiler do the copy of `bases` instead of copying it from a reference (big speedup here)
- Using `size_t` as the template parameter lets the compiler infer successfully
- Use `std::apply` instead of custom wrapper (using `std::tuple_cat`)
- Reordered the arguments to the callback to have the slice extent first.
- Move `make_for_each_contiguous_slice_dims` to cc file, with special cases.