@dsharlet 2024-02-13
This branch:
```
-----------------------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------
BM_for_each_contiguous_slice/64/16/1       27.7 ns         27.7 ns     25168431 bytes_per_second=66.6941Gi/s
BM_for_each_contiguous_slice/64/4/4        27.8 ns         27.8 ns     25310643 bytes_per_second=66.5012Gi/s
```
main:
```
-----------------------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------
BM_for_each_contiguous_slice/64/16/1       34.0 ns         34.0 ns     20629549 bytes_per_second=54.2998Gi/s
BM_for_each_contiguous_slice/64/4/4        33.9 ns         33.9 ns     20285857 bytes_per_second=54.5572Gi/s
```
18ef707
```
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_for_each_contiguous_slice/64/16/1       26.8 ns         26.8 ns     26601849
BM_for_each_contiguous_slice/64/4/4        25.4 ns         25.4 ns     26708943
```